+++ addons/eu/en16931/en16931.go
@@ -36,6 +36,7 @@ func init() {
 		payInstructionsRules(),
 		payTermsRules(),
 		orgItemRules(),
+		orgIdentityRules(),
 		orgAttachmentRules(),
 		orgPartyRules(),
 		orgInboxRules(),